+++ packages/sampler/stories/default/DropDown.js
@@ -1,4 +1,4 @@
-import DropDown, {DropDownBase} from '@enact/moonstone/DropDown';
+import Dropdown, {DropdownBase} from '@enact/moonstone/DropDown/Dropdown';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
@@ -7,19 +7,19 @@ import {withInfo} from '@storybook/addon-info';
 import {boolean, select, text} from '../../src/enact-knobs';
 import {mergeComponentMetadata} from '../../src/utils';
 
-const Config = mergeComponentMetadata('DropDown', DropDown, DropDownBase);
-DropDown.displayName = 'DropDown';
+const Config = mergeComponentMetadata('Dropdown', Dropdown, DropdownBase);
+Dropdown.displayName = 'Dropdown';
 
 
 
 storiesOf('Moonstone', module)
 	.add(
-		'DropDown',
+		'Dropdown',
 		withInfo({
-			text: 'Basic usage of DropDown'
+			text: 'Basic usage of Dropdown'
 		})(() => (
 			<div>
-				<DropDown
+				<Dropdown
 					direction={select('direction', ['up', 'down'], Config)}
 					disabled={boolean('disabled', Config)}
 					onClose={action('onClose')}
@@ -28,7 +28,7 @@ storiesOf('Moonstone', module)
 					title={text('title', Config, 'Dropdown')}
 				>
 					{['Option 1', 'Option 2', 'Option 3', 'Option 4', 'Option 5']}
-				</DropDown>
+				</Dropdown>
 			</div>
 		))
 	);